To fix integer handling coverity check defect

+++ src/upstream.h
@@ -31,7 +31,7 @@ void packMetaData();
 void *handle_upstream();
 void *processUpstreamMessage();
 void getParodusUrl();
-void sendUpstreamMsgToServer(void **resp_bytes, int resp_size);
+void sendUpstreamMsgToServer(void **resp_bytes, size_t resp_size);
 
 #ifdef __cplusplus
 }
